add fzf wrapper
as pretty as this little function is, shouldn't be stuck in a .zshrc.
adding more and more things like this will slow stuff down, and it's a
mess for portability too, with the whole .zshrc-linux .zshrc-freebsd
.zshrc-linux-desktop thing going on.

instead, going to move all these things to .config/init/funcs and add a
system which is called from .zprofile (and should move
.config/init/init.sh to being run there and .config/init/vars to being
sourced there too.). this system should check prereqs for each func and
load only when they're satisfied.